Refactor Github Actions (#264)
* Deprecate coverage action inputs

They are no longer used.

1. To generate code coverage reports latest available coverage package
   is installed as library by ST (requires Package Control!). 
   Coverage reports are created from within ST's plugin environment.
2. Coverage reports are to be uploaded via via codecov action, which has to
   deal with installing required coverage python packages itself.

* Drop obsolete package recovery

Package Control is configured not to auto-update or touch any package.
Restoring something after installing Package Control is obsolete.

* Add step names

* Don't start ST after installing

Starting and stopping ST is just a waste of time.
It will be started to install Package Control.

* Enable coverage on MacOS

* Deprecate install-unittesting action input

1. Allowing to not install UnitTesting is odd as tests won't run without it.
2. Installing a version which is unrelated with current action revision is
   also odd and confusing. If a certain version is desired for CI, it can be
   selected via `SublimeText/UnitTesting/actions/setup@<revision>`.

The `setup` action always installs the UnitTesting package, which is bundled
with the action's revision, if UnitTesting is not the checked out package or
hasn't been installed as extra-package. That's possible as the action is part
of UnitTesting package. The package doesn't need to be re-downloaded.

* Refactor setup steps

* On demand download Package Control

This commit prevents Package Control.sublime-package from being downloaded
and installed, if Package Control folder already exists as extracted package.

This may be the case if it is the package to test or was installed
via `extra-packages` input.

+++ actions/setup/install_package_control.sh
@@ -42,7 +42,7 @@ if [ ! -d "$STIP" ]; then
 fi
 
 PC_PATH="$STIP/Package Control.sublime-package"
-if [ ! -f "$PC_PATH" ]; then
+if [ ! -f "$PC_PATH" ] && [ ! -d "$STP/Package Control" ]; then
     echo Downloading Package Control.sublime-package
     PC_URL="https://github.com/wbond/package_control/releases/latest/download/Package.Control.sublime-package"
     curl -s -L "$PC_URL" -o "$PC_PATH"
